fix: 标签归并 SQLITE_CONSTRAINT_UNIQUE 崩溃、页面不刷新、颜色透明、映射后标签残留
- saveTag: update 失败时自动降级为归并流程，防御 getByName 返回 null 导致的 UNIQUE 约束崩溃
- loadDataBody: observeAll().first() 改用 getAllSync()，避免 Room Flow 共享缓存导致数据不刷新
- 颜色: Int→Long 符号扩展修复(改用 toUInt().toLong())，显示侧统一用 Color(Int) 构造函数
- upsertMapping/setStandard: 映射插入失败时 catch + delete+insert 重试，防止 mergeAliasTagInto 成功后映射丢失导致标签重建

+++ app/src/main/java/io/legado/app/ui/book/info/BookInfoScreen.kt
@@ -1031,7 +1031,7 @@ private fun BookInfoHeader(
                         key = { tag -> "colored-${tag.id}-${tag.name}" }
                     ) { tag ->
                         val color = if (tag.color != 0L) {
-                            Color(tag.color)
+                            Color(tag.color.toInt())
                         } else {
                             LegadoTheme.colorScheme.primary
                         }
